# perf(linker): Fixes brute force chunk cycle detection 