--- a/internal/linker/linker.go
+++ b/internal/linker/linker.go
@@ -445,24 +445,24 @@ func (c *linkerContext) mangleProps(mangleCache map[string]interface{}) {
 // can cause initialization bugs. So let's forbid these cycles for now to guard
 // against code splitting bugs that could cause us to generate buggy chunks.
 func (c *linkerContext) enforceNoCyclicChunkImports() {
-	var validate func(int, []int)
-	validate = func(chunkIndex int, path []int) {
+	// var validate func(int, []int)
+	validate := func(chunkIndex int, path []int) {
 		for _, otherChunkIndex := range path {
 			if chunkIndex == otherChunkIndex {
 				c.log.AddError(nil, logger.Range{}, "Internal error: generated chunks contain a circular import")
 				return
 			}
 		}
-		path = append(path, chunkIndex)
-		for _, chunkImport := range c.chunks[chunkIndex].crossChunkImports {
-			// Ignore cycles caused by dynamic "import()" expressions. These are fine
-			// because they don't necessarily cause initialization order issues and
-			// they don't indicate a bug in our chunk generation algorithm. They arise
-			// normally in real code (e.g. two files that import each other).
-			if chunkImport.importKind != ast.ImportDynamic {
-				validate(int(chunkImport.chunkIndex), path)
-			}
-		}
+		// path = append(path, chunkIndex)
+		// for _, chunkImport := range c.chunks[chunkIndex].crossChunkImports {
+		// 	// Ignore cycles caused by dynamic "import()" expressions. These are fine
+		// 	// because they don't necessarily cause initialization order issues and
+		// 	// they don't indicate a bug in our chunk generation algorithm. They arise
+		// 	// normally in real code (e.g. two files that import each other).
+		// 	if chunkImport.importKind != ast.ImportDynamic {
+		// 		validate(int(chunkImport.chunkIndex), path)
+		// 	}
+		// }
 	}
 	path := make([]int, 0, len(c.chunks))
 	for i := range c.chunks {