@evanw 2023-04-22
Thanks for the improvement and for the detailed report.

However, note that encountering this problem likely means that esbuild is a poorly-suited tool for your use case. It's not good to have too many chunks at run-time and esbuild doesn't currently have the ability to control the number of chunks (as it's code splitting feature is still in a WIP state). So you probably want to look for another tool even after this improvement has landed.