@MLoughry 2023-04-22
@evanw We currently use `esbuild` for dev innerloop builds, where the output is "good enough" and the build speed is much faster than our `webpack` dev builds. But we will certainly keep that in mind if/when we investigate bundlers aside from `webpack` for production builds.